utils/bundle_miner: fix type mismatch

+++ utils/bundle_miner.c
@@ -61,7 +61,7 @@ static void *bundle_miner_mine_routine(void *const param) {
       if (probability < ctx->probability) {
         ctx->probability = probability;
         ctx->optimal_index = ctx->index;
-        if (1 / (probability) >= ctx->mining_threshold) {
+        if ((uint32_t)(1 / (probability)) >= ctx->mining_threshold) {
           break;
         }
       }
@@ -103,7 +103,7 @@ void bundle_miner_normalized_bundle_max(byte_t const *const lhs, byte_t const *c
 
 retcode_t bundle_miner_mine(byte_t const *const bundle_normalized_max, uint8_t const security,
                             trit_t const *const essence, size_t const essence_length, uint32_t const count,
-                            uint8_t const nprocs, uint64_t mining_threshold, uint64_t *const index) {
+                            uint8_t const nprocs, uint32_t mining_threshold, uint64_t *const index) {
   retcode_t ret = RC_OK;
   uint8_t procs = nprocs == 0 ? system_cpu_available() : nprocs;
   uint32_t rounded_count = count + (procs - count % procs);
