+++ utils/bundle_miner.h
@@ -55,7 +55,7 @@ void bundle_miner_normalized_bundle_max(byte_t const *const lhs, byte_t const *c
  */
 retcode_t bundle_miner_mine(byte_t const *const bundle_normalized_max, uint8_t const security,
                             trit_t const *const essence, size_t const essence_length, uint32_t const count,
-                            uint8_t const nprocs, uint64_t mining_threshold, uint64_t *const index);
+                            uint8_t const nprocs, uint32_t mining_threshold, uint64_t *const index);
 
 #ifdef __cplusplus
 }
